+++ lib/flutter_animate.dart
@@ -1,5 +1,6 @@
 export 'src/animate.dart';
 export 'src/animate_list.dart';
+export 'src/effect_list.dart';
 export 'src/flutter_animate.dart';
 export 'src/adapters/adapters.dart';
 export 'src/effects/effects.dart';